pyproject.toml
[tool.poetry]
name = "diagnosys-api"
version = "0.1.0"
description = ""
authors = ["Shav Vimalendiran <shavgraphics@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
llama-index = "^0.9.17"
pypdf = "^3.17.0"
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"
google-generativeai="^0.3.0"
qdrant_client="^1.7.0"
matplotlib="3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
